[simple-memo] refactor : DTO의 비즈니스 로직을 분리해서 VO로 생성
